add a loading state
