+++ app/assets/stylesheets/application.css
@@ -88,7 +88,7 @@ body {
 }
 
 .pie-chart {
-  width: 5rem;
+  width: 7rem;
 }
 
 .button-footer {
