+++ app/javascript/constants/actionTypes.js
@@ -1,6 +1,7 @@
 export const UPDATE_DATA = 'UPDATE_DATA';
 export const LOGIN = 'LOGIN';
 export const LOGOUT = 'LOGOUT';
+export const LOADING = 'LOADING';
 export const FEED_STACKS = 'FEED_STACKS';
 export const CREATE_STACK = 'CREATE_STACK';
 export const FEED_PROGRESS = 'FEED_PROGRESS';
